+++ BAG.CommandQL/ResponseQL.cs
@@ -22,7 +22,7 @@ namespace BAG.CommandQL
         public List<string> Errors { get; set; }
 
         //Time-Durration ms
-        public int T { get; set; }
+        public long T { get; set; }
     }
 
     //[JsonArray]
